MAINT: Fix build warning ([-Wpointer-to-int-cast])

+++ scipy/sparse/linalg/_dsolve/SuperLU/SRC/smemory.c
@@ -646,8 +646,8 @@ sStackCompress(GlobalLU_t *Glu)
 
     last = (char*)usub + xusub[ndim] * iword;
     fragment = (char*) (((char*)Glu->stack.array + Glu->stack.top1) - last);
-    Glu->stack.used -= (long int) fragment;
-    Glu->stack.top1 -= (long int) fragment;
+    Glu->stack.used -= (uintptr_t) fragment;
+    Glu->stack.top1 -= (uintptr_t) fragment;
 
     Glu->ucol = ucol;
     Glu->lsub = lsub;
